+++ src/imaging/opencv_camera.cc
@@ -50,7 +50,8 @@ void tfv::OpenCvUSBCamera::close(void) {
     camera_ = nullptr;
 }
 
-void tfv::OpenCvUSBCamera::retrieve_properties(size_t& width, size_t& height,
+void tfv::OpenCvUSBCamera::retrieve_properties(uint16_t& width,
+                                               uint16_t& height,
                                                size_t& frame_bytesize) {
 
     _retrieve_properties();
